+++ doc/changelog.md
@@ -13,6 +13,7 @@ Jump to:
 
 Description
 
+- Update SmartSim environment variables using new naming convention
 - Refactor `exception_handler`
 - Add RequestDispatcher and the possibility of batching inference requests
 - Enable hostname selection for dragon tasks